refactor: rename campus.client.base to campus.client.wrapper for clarity
